move code to node_api_embedding.cc
